Backend/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
audioread==3.0.1
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
Flask==3.1.0
Flask-SocketIO==5.5.0
flatbuffers==24.12.23
gast==0.6.0
gevent==24.11.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.7.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.1.0
namex==0.0.8
numba==0.60.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.2
pycparser==2.22
Pygments==2.18.0
python-engineio==4.11.1
python-socketio==5.12.0
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
simple-websocket==1.1.0
six==1.17.0
soundfile==0.12.1
soxr==0.5.0.post1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
